Refactored organizer autocomplete, so now it's not using formValues
