+++ EventsExpress/ClientApp/src/components/event/filter/parts/organizer/organizer-filter.js
@@ -1,7 +1,6 @@
 import { FilterExpansionPanel } from '../../expansion-panel/filter-expansion-panel';
 import { change, Field, getFormValues } from 'redux-form';
 import React, { useEffect } from 'react';
-import { Chip } from '@material-ui/core';
 import { useOrganizerFilterStyles } from './organizer-filter-styles';
 import { connect } from 'react-redux';
 import OrganizerAutocomplete from './organizer-autocomplete';
@@ -9,13 +8,6 @@ import { fetchUsers } from '../../../../../actions/events/filter/users-data';
 
 const OrganizerFilter = ({ organizers, formValues, fetchUsers, change }) => {
     const classes = useOrganizerFilterStyles();
-
-    const deleteOrganizer = organizerToDelete => {
-        return () => {
-            change(formValues.organizers.filter(organizer => organizer.id !== organizerToDelete.id));
-        };
-    };
-
     const clear = () => change([]);
 
     useEffect(() => {
@@ -34,16 +26,6 @@ const OrganizerFilter = ({ organizers, formValues, fetchUsers, change }) => {
                     options={organizers}
                     component={OrganizerAutocomplete}
                 />
-                <div className={classes.chips}>
-                    {formValues.organizers.map(organizer => (
-                        <Chip
-                            key={organizer.id}
-                            label={organizer.username}
-                            color="secondary"
-                            onDelete={deleteOrganizer(organizer)}
-                        />
-                    ))}
-                </div>
             </div>
         </FilterExpansionPanel>
     );
